+++ corehq/apps/data_cleaning/tests/test_views.py
@@ -25,7 +25,7 @@ from corehq.apps.data_cleaning.views.tables import (
     CleanCasesTableView,
     CaseCleaningTasksTableView,
 )
-from corehq.apps.data_cleaning.views.setup import (
+from corehq.apps.data_cleaning.views.start import (
     StartCaseSessionView,
 )
 from corehq.apps.domain.shortcuts import create_domain
