+++ corehq/apps/data_cleaning/views/main.py
@@ -33,7 +33,7 @@ class BulkEditCasesMainView(BaseProjectDataView):
 
     @property
     def page_context(self):
-        from corehq.apps.data_cleaning.views.setup import StartCaseSessionView
+        from corehq.apps.data_cleaning.views.start import StartCaseSessionView
         from corehq.apps.data_cleaning.views.tables import CaseCleaningTasksTableView
         return {
             "htmx_start_session_form_view": reverse(StartCaseSessionView.urlname, args=(self.domain,)),
